Устранение ошибок #3
