+++ Navigation/Resurses/Assets.xcassets/AppIcon.appiconset/Contents.json
