+++ Navigation/Resurses/Assets.xcassets/myfoto.imageset/Contents.json
