+++ Navigation/Resurses/Assets.xcassets/myfoto.imageset/myfoto.jpg
